(@bug W-3169552@) Read updated LAC unconditionally in proxy
If user is trying to ready beyond current LAC go ahead and call lh.readLac
this updates current LAC, and if the ledger is closed it updates from local metadata
hence no need to worry about protocol messages between client and Bookie.

Also fixed ReadOnlyLedgerHandle to update LAC and length of the ledger on metadata
refresh.

Signed-off-by: Venkateswararao Jujjuri (JV) <vjujjuri@salesforce.com>