(@bug W-3198184@) Allow recovery reads to fence open writers within same proxy
The ledgerRecoveryOpenRead implementation in bkproxy suppresses fencing reads
for ledgers open for write access, under the expectation that that an sdb
server would never do such a thing.  However, we do need to be able to do this
for log fencing when promoting HA and DR standby servers, and will definitely
have deployment scenarios where multiple sdb servers could be using the same
proxy instance.  To protect against split-brain, we need the fencing requests
to always make it to the underlying bookkeeper cluster.

This change is a small patch to have ledgerRecoveryOpenRead close any open write
handle and arrange for the fencing read to go to bookkeeper.