(@bug W-3013711@) Sfstore changes for creating a generic (K,V) map for storing ledger metadata
Create a (K,V) map for storing ledger metadata in zoo keeper. For example, can be used for storing creatorId of a ledger.

This also has:

(i) Changes to introduce a protocol version field in the messages between sayonara->proxy.
(ii) Change the ledgerList interface to return the list in batches since that reduces the no. of times we need to scan the list from zookeeper.