(@bug W-3315975@) Provide an option to delay auto recovery of lost bookies for planned downtimes or intermittent
When auto recovery is enabled, if a node goes down for upgrade or even if it looses zk connection
intermittently, the auditor detects it as lost bookie and starts under replication detection and
the replication workers on bookie nodes start replicating under replicated ledgers. All of this
stops once the bookie comes up but by then a few ledgers would get replicated. Given the fact that
we have 3 copies of data, it is probably not necessary to start the recovery as soon as a bookie
goes down. We can probably wait for an hour or so and then start recovery. This should cover cases
like planned upgrade or a network glitch.