(@bug W-3315975@) Provide an option to delay auto recovery of lost bookies for planned downtimes or intermittent
Fixing a bug in the earlier implementation where, on a rolling upgrade, we were not honoring 'lostBookieRecoveryDelay'.
I.e. if the first bookie goes down, we honor 'lostBookieRecoveryDelay'. But if that bookie comes up and a different
bookie goes down we would treat it like multiple bookie failure and start the audit. This change fixes this problem
such that when the second bookie goes down it realizes that there is only one bookie down and hence doesn't treat
it like a multiple bookie failure scenario.