(@bug W-3379495@) Use shutDownNow instead shutdown
- with the new approach we are scheduling new tasks for
CHECK_ALL_LEDGERS and BOOKIE_CHECK with delay. When 'shutdown'
of ExecutorService is invoked the tasks which are never commenced
are not cancelled/halted. So in Auditor.shutdown method it awaits
for full 30 secs and the testcases are failing with timeouts or with
failure checks.

Previously we use to submit tasks with 'scheduleWithFixedDelay'
(periodic action), and if shutdown is called in the period
between successive executions then it was getting cancelled, so
it used to work correctly.

It is ok to call shutdownnow instead of shutdown, because we are
submitting this shutdownnow task to 'executor'
(singleThreadScheduledExecutor), so if it is actively any task
(BookieCheck or checkAllLedgers) then this will be called after
completion of that task.