BOOKKEEPER-946: Provide an option to delay auto recovery of lost bookies
***Duplicate of c0ca28fc752c5deccec4c985ded043770b1e1f80/W-3315975

If auto recovery is enabled, and a bookie goes down for upgrade or even if it looses zk connection
intermittently, the auditor detects it as a lost bookie and starts under replication detection and
the replication workers on other bookie nodes start replicating the under replicated ledgers. All
of this stops once the bookie comes up but by then a few ledgers would get replicated. Given the
fact that we have multiple copies of data, it is probably not necessary to start the recovery as
soon as a bookie goes down. We can wait for an hour or so and then start recovery. This should
cover cases like planned upgrade, intermittent network connectivity loss, etc.

This change:
    1) Provides a bookie option 'lostBookieRecoveryDelay' in secs, which when set to a non zero value,
       will delay the start of recovery by that number of seconds. By default, this option is set to 0;
       which means the audit is not delayed.
    2) If another bookie were to go down in this interval, the recovery is immediately started and the
       one scheduled for future is canceled.
    3) Adds counters to track how many audits were delayed(#1) and how many scheduled audits were
       canceled due to multiple bookie failures(#2).
    4) Three junit tests to verify the new feature.

Author: Rithin <rithin.shetty@salesforce.com>

Reviewers: siddharth.boobna@gmail.com <siddharth.boobna@gmail.com>, Enrico Olivelli <eolivelli@gmail.com>

Closes #58 from rithin-shetty/audit_delay