(@bug W-3417165@) Enable rackaware ensemble placement policy
This changes enables rackaware ensemble placement policy by providing a python
script for resolving the bookie IP addresses to hostnames. The script tries to
find the data center in which it is running using the hostname of the bk client
machine. It then uses that information to read the racks->bookieIP mapping from
the corresponding yaml file. For now, the yaml files will be part of the bkproxy
package. We will need to revisit this in the future when the number of DCs in which
sfstore is running goes up. If the yaml file is missing, rackaware ensemble policy
will treat all the bookies to be on the same rack and will randomly select bookies
from it. The same is true if an entry is missing for a bookie in the yaml file: it
will be treated as beloging to default rack and will be selected for ledger allocation.