(@bug W-3693842@) Fix in logic of GCThread logic
- when it is running forcegc, forceGarbageCollection
should be set to false after it has done major/minor
compaction (continue statement should be removed).
Finally should be used.
- it is optimal to do minorcompaction first, then
majorcompaction in the case of forcegc. Since it makes
bookie to reclaim space faster (if available) and turn
it to RW when LedgerDirsMonitor calculates the disk
usage and bring up the writableLedgerDirs.