(@bug W-3198146@) report request stats based on send completion
Previously, addRequestStats and readRequestStats reflected the time from
receipt of the request through to submitting the response to the
channel.  This patch expands the scope through to where the channel
actually does the send in order to correctly reflect situations where
responses are blocked on sending for whatever reason.

Additionally, introduce a single counter to count time required to
perform the actual send.

Signed-off-by: Samuel Just <sjust@salesforce.com>