(@bug W-3962128@) Entrylog per ledger
Current bookkeeper is tuned for rotational HDDs. It has one active entrylog,
and all the ledger/entries go to the same entrylog until it is rotated out.
This is perfect for HDDs as seeks and moving head allover the disk platter
is very expensive. But this is very inefficient for SSDs, as each SSD
can handle multiple parallel writers, also this method is extremely
inefficient for compaction as it causes write amplification
and inefficient disk space usage.

Our proposal is to have configurable feature to have entrylog per ledger
and also to have parallel EntryMemTable flusher for SortedLedgerStorage.

Implementation Details

- created Interface and corresponding implementations for EntryLogManager
- changed logic of EntryMemTable flush parallely to have just one Runnable per ledger
- moved the complete logic of flushIntervalInBytes from EntryLogger to BufferedChannel
- added more testcases