BOOKKEEPER-1041: Multiple ActiveEntrylogs
Current bookkeeper is tuned for rotational HDDs. It has one active entrylog,
and all the ledger/entries go to the same entrylog until it is rotated out.
This is perfect for HDDs as seeks and moving head allover the disk platter
is very expensive. But this is very inefficient for SSDs, as each SSD
can handle multiple parallel writers, also this method is extremely
inefficient for compaction as it causes write amplification
and inefficient disk space usage.

Our proposal is to have multiple active entrylogs and a configuration
param on how many parallel entrylogs per ledgerdir can have.
This way one can have ability to configure to have less
(may be one) ledgers per entrylog.