(@bug W-4268290@) Enable stats on OrderedSafeExecutor
It costs one additional allocation per item, but we
want the stats.  If the allocation is a problem, we'll
add an object pool.

Fix taskPendingStats value in TimedRunnable

There don't appear to be any users at this time with the default
configuration, so it was probably just at typo.

BookieRequestProcessor: clarify thread pool names

No need to include the port numbers in the thread pool names since there
is only one of each in this metric namespace and log.  The
OrderedSafeExecutor implementation already introduces further namespaces
to disambiguate the different threads.

Enable op stat tracing on Bookie read/write tps

Introduces a config as on the client to enable fine-grained operation
latency tracking and passes the bookkeeper_server statLogger into the
thread pools.

@rev ayegorov@
@rev vjujjuri@
Signed-off-by: Samuel Just <sjust@salesforce.com>