(@bug W-4276826@) Add bookie read/write stat sanity checks
Adds a trivial mock for remembering stats long enough to verify that
counters are actually used and sensible in unit tests and bakes it into
BookKeeperClusterTestCase so that we can write tests to ensure that the
stats are actually counted and make sense.

Uses above to simple tests for top level read and write stats.

@rev ayegorov@
@rev vjujjuri@
Signed-off-by: Samuel Just <sjust@salesforce.com>