(@bug W-4709063@) (#78)
If log4j is configured for asynchronous logging, when the jar
execution completes gracefully, the shutdown hook on log4j is
triggered but fails to flush the remaining log events. This seems to
be a bug in Log4j2.

https://issues.apache.org/jira/browse/LOG4J2-658

Here, disabling asynchronous logging only on bookie shell commands.
Bookie and Localbookies continue to use asynchronous logging.