AbstractConfiguration: do not prepend cluster name on setProperty
This feature is actually broken is several ways.  One of them is that
loadConf now uses setProperty to set the keys.  That definitely doesn't
work with the magic property prepending.  To mitigate that issue (though
incompletely), setProperty will set an unadorned value of that name.
That means that it won't override cluster specific settings.  It also
means that you don't get cluster specific setting unless you do it on
purpose.  I claim that that is actually more predictable behavior
anyway.

This needs another look prior to switching over TODO.

Signed-off-by: Samuel Just <sjust@salesforce.com>