(@bug W-4698540@) Increase openLedgerRereplicationGracePeriod (#77)
* (@bug W-4698540@) Increase openLedgerRereplicationGracePeriod

openLedgerRereplicationGracePeriod defines how long the replication
worker waits to fence an active extent before starting replication.
The default value is 30 sec. Which should be enough for the writer
to detect the bookie down incident and make an ensemble change.
But given that we don't have back-pressure, sometimes client and
bookies get backed up and may take longer time to detect and change
the ensemble. Changing this value to 10 mins.

Making this longer should not have any major impact as the writer is
confirmed to everything it wrote. Also now we have explicitLAC so
even delaying the fencing won't stop/restrict reader reading even
the last entry/fragment

Signed-off-by: Venkateswararao Jujjuri (JV) <vjujjuri@salesforce.com>

* Applied review comments