(@bug W-4812279@) If SSL handshake fails, channel close can cause the netty IO thread to hang indefinitely
SSL handshake failure is handled in the netty IO thread context. Failure
in handshake results in channel close and waiting for it indefinitely.
Upon channel close, the notification is also processed in the netty IO
thread context, which can cause deadlock.

In this specific case of handling handshake failure, waiting for channel
close is not necessary.

If wait is necessary, then it should be handled outside the netty IO
thread context.

Make close() operation default to not wait.

@rev ayegorov@