(@W-4963336@) Adding mechanism to publish an eRelease build for fixes (#123)
Modified script to check if the current release being targeted is a normal
or an eRelease. In case of an eRelease, increment the patch version while
retaining the older minor version. In case if weekly release, create a new 
branch with the commits for the release and push the tag.

(@rev kudayashankar@)
(@rev sjust@)

Signed-off-by: Ankit Jain <jain.a@salesforce.com>