(@bug W-5144037@) Improve bookkeeper build time for SDB (#134)
@sjust 3938895

Sadly, it's not possible (afaict) for a profile to reduce the module set
from the build section. Instead, create a default profile with the full
set and non-default minimal profile without it.

This patch also removes the extra upstream profiles because they become
brittle to maintain since apparently I can't chain them to the default
one. Since we aren't using them, safer to get rid of them.

Signed-off-by: Samuel Just sjust@salesforce.com

@jain-a d9af6cf

Disable circe-checksum tests when building sfstore.

Signed-off-by: Ankit Jain jain.a@salesforce.com