(@bug W-5357523@) k8sstartup: handle lines with key = value properly
k8sstartup assumed that all lines had the form

key=value

however

key = value

is also valid for the config file.  As it happens,
openLedgerRereplicationGracePeriod is in that form, so k8sstartup failed
to override it.  Update k8sstartup to strip the extra spacees before
matching the key.

(@rev dcastor@)
(@rev vjujjuri@)
Signed-off-by: Samuel Just <sjust@salesforce.com>