autorecovery-use-metadata-driver (part 1) : move AutoRecoveryMain to use MetadataBookieDriver
Descriptions of the changes in this PR:

We are introducing Etcd as the metadata storage. However AutoRecovery currently is still tied to zookeeper. In order to use Etcd as the metadata storage, we have to move AutoRecovery related classes to use metadata driver API.

This is the first change for changing AutoRecovery to use metadata driver. It changed AutoRecoveryMain to use metadata driver api and also removed the shutdown logic on session expired, which doesn't make any sense for current retryable zookeeper.

Author: Qi Wang <codingwangqi@gmail.com>

Reviewers: Sijie Guo <sijie@apache.org>, Enrico Olivelli <eolivelli@gmail.com>

This closes #1693 from codingwangqi/autorecovery_metadata